+++ task2/frontend/src/environments/environment.ts
@@ -1,5 +1,5 @@
 export const environment = {
   production: false,
-  maxTrials: 3,
+  maxTrials: 6,
   apiHost: 'http://localhost:5043'
 };